made seperate command for vision drive in auto
Made a new command called VisionDriveAuto for VDrive during autonomous. It is the same as VisionDrive, but will end automatically after a certain amount of time or after the error is less than a certain value. The normal VisionDrive only ends if interrupted, and should not be used in sequential command groups during auto.